+++ CHANGELOG.md
@@ -4,6 +4,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v1.29.0
+------------------------------
+*September 14, 2018*
+
+### Added
+- Scroll spy module.
+- Lodash throttle module.
+
+### Changed
+- Updated menu layouts.
+
+
 v1.28.0
 ------------------------------
 *September 12, 2018*